feat(linked-list): Make the list own its data
Up until this point, `LinkedList` accepted references for each element,
instead of owning them.

Having read-only references does not allow `LinkedList` to have update
operations or anything else that requires mutable borrows or owned data
(unless T is Clone).

Therefore, this commit changes the data structure to own it's data,
which allows it to be a lot more flexible.

Signed-off-by: acikgozb <acikgozb@proton.me>